libs/vizu_google_suite_client/pyproject.toml
[tool.poetry]
name = "vizu_google_suite_client"
version = "0.1.0"
description = "Google Suite client (Sheets, Gmail, Calendar) for vizu-mono"
authors = ["Vizu <dev@vizu.example>"]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.110.0"
google-auth = "^2.25.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
pydantic = "^2.5.0"
httpx = ">=0.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
